Allow specifying an alternate method of getting a raw body
This can be used if the raw body is located elsewhere in the conn.

In our situation we already are pattern matching on the path in the
Phoenix endpoint module with the util added in the test which came
from the
[Bytepack archive](https://github.com/dashbitco/bytepack_archive/blob/main/apps/bytepack_web/lib/bytepack_web/extensions/plug/body_reader.ex),
so this felt much cleaner to support in the plug than define another
alternate route that stores raw bodies slightly different.